requirements.txt
Pillow==10.3.0
numpy==1.26.4
periphery==2.1.1
spidev==3.6